Move groupingStrInRange to new ValidationUtils file
