+++ qt/scientific_interfaces/Inelastic/Common/InterfaceUtils.h
@@ -41,9 +41,6 @@ MANTIDQT_INELASTIC_DLL QStringList getCorrectionsWSSuffixes(std::string const &i
 MANTIDQT_INELASTIC_DLL QPair<double, double> convertTupleToQPair(std::tuple<double, double> const &doubleTuple);
 MANTIDQT_INELASTIC_DLL std::pair<double, double> convertTupleToPair(std::tuple<double, double> const &doubleTuple);
 MANTIDQT_INELASTIC_DLL QString makeQStringNumber(double value, int precision);
-
-MANTIDQT_INELASTIC_DLL bool groupingStrInRange(std::string const &customString, std::size_t const &spectraMin,
-                                               std::size_t const &spectraMax);
 } // namespace InterfaceUtils
 } // namespace CustomInterfaces
 } // namespace MantidQt
